Edits to ltcv.py, get ltcvapp.py working again

+++ tests/admin/test_load_snana_fits.py
@@ -46,10 +46,12 @@ def test_load_snana_fits():
             for tab in [ 'host_galaxy', 'diaobject', 'diasource', 'diaforcedsource', 'root_diaobject' ]:
                 cursor.execute( f"TRUNCATE TABLE {tab} CASCADE" )
             cursor.execute( "SELECT id FROM processing_version WHERE description='test_procver'" )
-            pid = cursor.fetchone()[0]
-            cursor.execute( "DELETE FROM base_procver_of_procver WHERE procver_id=%(pv)s", { 'pv': pid } )
-            cursor.execute( "DELETE FROM processing_version WHERE description='test_procver'" )
-            cursor.execute( "DELETE FROM base_processing_version WHERE description='test_procver'" )
+            pid = cursor.fetchone()
+            if pid is not None:
+                pid = pid[0]
+                cursor.execute( "DELETE FROM base_procver_of_procver WHERE procver_id=%(pv)s", { 'pv': pid } )
+                cursor.execute( "DELETE FROM processing_version WHERE description='test_procver'" )
+                cursor.execute( "DELETE FROM base_processing_version WHERE description='test_procver'" )
             conn.commit()
